develop child comment count update to redis and select form redis

+++ comment/comment-biz/src/main/java/com/danby/happynode/comment/biz/domain/dataobject/CommentDO.java
@@ -24,7 +24,7 @@ public class CommentDO {
 
     private String imageUrl;
 
-    private Byte level;
+    private Integer level;
 
     private Long replyTotal;
